Make sure the JVM actually has to work for a bit
This should be enough to ensure that JIT compiler actually kicks in, and
that the JVM has to allocate some memory on behalf of the program. (I
thought that last part might indicate that we’d need to add brk(2) to
the system call filter, but apparently not.)

(Also expand the commented-out debugging SystemCallFilter, which I’ll
remove in a subsequent commit. If you want to use it later, don’t forget
to also remove the line in javad.service.d/187.conf that clears the
filter before adding to it.)